added sample for code first grpc and documentation update (#22128)
