--- aspnetcore/grpc/code-first/Contracts.cs
@@ -1,21 +1,0 @@
-[DataContract]
-public class HelloRequest
-{
-    [DataMember(Order = 1)]
-    public string Name { get; set; }
-}
-
-[DataContract]
-public class HelloReply
-{
-    [DataMember(Order = 1)]
-    public string Message { get; set; }
-}
-
-[ServiceContract]
-public interface IGreeterService
-{
-    [OperationContract]
-    Task<HelloReply> SayHelloAsync(HelloRequest request, 
-        CallContext context = default);
-}
