--- aspnetcore/grpc/code-first/GreeterService.cs
@@ -1,11 +1,0 @@
-public class GreeterService : IGreeterService
-{
-    public Task<HelloReply> SayHelloAsync(HelloRequest request, 
-        CallContext context = default)
-    {
-        return Task.FromResult(
-            new HelloReply {
-                Message = $"Hello {request.Name}"
-            });
-    }
-}
