--- aspnetcore/grpc/code-first/Startup.cs
@@ -1,19 +1,0 @@
-public void ConfigureServices(IServiceCollection services)
-{
-    services.AddCodeFirstGrpc();
-}
-
-public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-{
-    if (env.IsDevelopment())
-    {
-        app.UseDeveloperExceptionPage();
-    }
-
-    app.UseRouting();
-
-    app.UseEndpoints(endpoints =>
-    {
-        endpoints.MapGrpcService<GreeterService>();
-    });
-}
