+++ aspnetcore/grpc/code-first/samples/5.x/GrpcGreeter/Startup.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using ProtoBuf.Grpc.Server;
+
+namespace GrpcGreeter
+{
+    public class Startup
+    {
+        #region snippet
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddCodeFirstGrpc();
+        }
+
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapGrpcService<GreeterService>();
+            });
+        }
+        #endregion
+    }
+}
