+++ aspnetcore/grpc/code-first/samples/5.x/Shared/Contracts.cs
@@ -1,0 +1,31 @@
+﻿using ProtoBuf.Grpc;
+using System.Runtime.Serialization;
+using System.ServiceModel;
+using System.Threading.Tasks;
+
+namespace Shared.Contracts
+{
+    #region snippet
+    [DataContract]
+    public class HelloReply
+    {
+        [DataMember(Order = 1)]
+        public string Message { get; set; }
+    }
+
+    [DataContract]
+    public class HelloRequest
+    {
+        [DataMember(Order = 1)]
+        public string Name { get; set; }
+    }
+
+    [ServiceContract]
+    public interface IGreeterService
+    {
+        [OperationContract]
+        Task<HelloReply> SayHelloAsync(HelloRequest request,
+            CallContext context = default);
+    }
+    #endregion
+}
